SFR-831 changed streamed example to only use NYPL reader
