+++ examples/static/static-server.js
@@ -21,7 +21,7 @@ server.listen( port, (err) => {
   }
 });
 
-app.use("/viewer", express.static(__dirname + "/viewer"));
+app.use("/viewer", express.static(__dirname + "/../../viewer"));
 app.use("/TheCallOfTheWild", express.static(__dirname + "/TheCallOfTheWild"));
 app.use("/AJourneyToTheCentreOfTheEarth", express.static(__dirname + "/AJourneyToTheCentreOfTheEarth"));
 app.use("/backstop", express.static(__dirname + "/backstop"));
