+++ examples/streamed/views/index.html.ejs
@@ -13,7 +13,7 @@
 </head>
 
 <body>
-  <h1>Publication URL aaaaaa</h1>
+  <h1>External Epub URL</h1>
   <form id="form">
     <input type="text" name="url" id="url" size="80">
     <input type="submit" value="Go!">
