--- examples/test/package.json
@@ -1,26 +1,0 @@
-
-{
-    "name": "webpub-viewer-demo",
-    "version": "0.0.2",
-    "description": "Demo server using the OPDS web client",
-    "repository": {
-      "type": "git",
-      "url": "https://github.com/NYPL-Simplified/opds-web-client.git"
-    },
-    "main": "index.js",
-    "author": "NYPL",
-    "license": "Apache-2.0",
-    "scripts": {
-      "dev": "NODE_ENV=development node index.js",
-      "start": "NODE_ENV=production node index.js"
-    },
-    "dependencies": {
-      "ejs": "^2.4.1",
-      "express": "^4.13.4",
-      "multer": "^1.1.0",
-      "nypl-simplified-webpub-viewer": "git+https://git@github.com/NYPL-Simplified/webpub-viewer#develop",
-      "r2-streamer-js": "0.0.1-alpha13",
-      "request": "^2.69.0",
-      "requirejs": "^2.3.2"
-    }
-  }
