+++ viewer/index.html
@@ -18,7 +18,6 @@
     var getURLQueryParams = function() {
       var params = {};
       var query = window.location.search;
-      console.log("query", query);
       if (query && query.length) {
         query = query.substring(1);
         var keyParams = query.split('&');
